+++ src/frontend/swiftui/Api.swift
@@ -232,7 +232,7 @@ struct HydraStringArrayOption {
     }
 
     func resize(newCount: Int) {
-        hydra_string_array_option_resize(self.handle, UInt64(newCount))
+        hydra_string_array_option_resize(self.handle, UInt32(newCount))
     }
 
     func set(at index: Int, value: String) {
